Move Handler related code to dummy_handler
